+++ SHARE.py
@@ -11,7 +11,7 @@ class Share:
 
         #share 화면 이미지
         self.shareBack = tkinter.PhotoImage(file="img/background.gif")
-        self.shareBackL = tkinter.Label(img=self.shareBack)
+        self.shareBackL = tkinter.Label(image=self.shareBack)
         self.shareBackL.place(x=-2, y=-2)
